tracing: add lock_contention_analysis.bt based on the sync traceponts
This bpftrace script can be used to analyze lock contention. It shows
code locations that have to wait a long time to actually acquire a
lock, and also shows the last location that lock was held, to see
who was actually responsible for blocking.

When monitoring ends, prints the locations with the longest waiting
durations.